@@ -1,63 +1,168 @@
+# Copyright 2011-2014 Rice University. Licensed under the Affero General Public 
+# License version 3 or later.  See the COPYRIGHT file for details.
+
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.17'
+# Rails framework
+gem 'rails', '4.2.0.beta2'
 
-# Gems used only for assets and not required
-# in production environments by default.
-group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
-  gem 'bootstrap-sass', '~> 3.1.1'
-  gem 'therubyracer', :platforms => :ruby
-  gem 'uglifier', '>= 1.0.3'
-  gem 'compass-rails'
-end
+# Remove this when Squeel is updated to support Arel 6.0.0
+gem 'arel', '6.0.0.beta2'
 
+# Bootstrap front-end framework
+gem 'bootstrap-sass', '~> 3.2.0'
+
+# SCSS stylesheets
+gem 'sass-rails', '~> 5.0.0.beta1'
+
+# Compass stylesheets
+gem 'compass-rails'
+
+# Automatically add browser-specific CSS prefixes
+gem 'autoprefixer-rails'
+
+# JavaScript asset compressor
+gem 'uglifier', '>= 1.3.0'
+
+# CoffeeScript for .js.coffee assets and views
+gem 'coffee-rails', '~> 4.0.0'
+
+# JavaScript asset compiler
+gem 'therubyracer', platforms: :ruby
+
+# jQuery library
 gem 'jquery-rails'
+
+# jQuery UI library
 gem 'jquery-ui-rails'
-gem 'jbuilder'
-gem 'whenever', :require => false
-gem 'newrelic_rpm'
-gem 'squeel'
-gem 'yaml_db'
 
-gem 'openstax_utilities', '~> 2.2.3'
-gem 'openstax_accounts', '~> 2.0.0'
-gem 'openstax_api', '~> 2.2.3'
+# Automatically ajaxify links
+gem 'turbolinks'
 
-gem 'apipie-rails'
-gem 'maruku'
+# Rails 5 HTML sanitizer
+gem 'rails-html-sanitizer', '~> 1.0'
 
+# Utilities for OpenStax websites
+gem 'openstax_utilities'
+
+# Cron job scheduling
+gem 'whenever'
+
+# OpenStax Accounts integration
+gem 'openstax_accounts', path: '../accounts-rails'
+
+# Respond_with and respond_to methods
+gem 'responders', '~> 2.0'
+
+# Access control for API's
 gem 'doorkeeper'
 
+# API versioning and documentation
+gem 'openstax_api'
+gem 'apipie-rails'
+gem 'maruku'
 gem 'representable'
 gem 'roar-rails'
 
-gem 'exception_notification'
+# Lev framework
+gem 'lev'
 
-gem 'action_interceptor', '~> 0.2.3'
-gem 'fine_print', '~> 1.4.1'
+# Ruby dsl for SQL queries
+gem 'squeel'
 
-gem 'lev', '~> 2.1.0'
+# Contract management
+gem 'fine_print', path: '../fine_print'
+
+# Keyword search
+gem 'keyword_search'
+
+# File uploads
+gem 'remotipart'
+gem 'carrierwave'
+
+# Image editing
+gem 'mini_magick'
+
+# Embedded JavaScript templates
+gem 'ejs'
+
+# Embedded CoffeeScript templates
+gem 'eco'
+
+# Real time application monitoring
+gem 'newrelic_rpm'
+
+# YAML database backups
+gem 'yaml_db'
 
 group :development, :test do
-  gem 'sqlite3'
-  gem 'debugger'
-  gem 'faker'
+  # Thin webserver
   gem 'thin'
+
+  # SQLite3 database
+  gem 'sqlite3'
+
+  # Call 'debugger' anywhere in the code to stop execution and get a debugger console
+  gem 'byebug'
+
+  # Mute asset pipeline log messages
   gem 'quiet_assets'
-  gem 'cheat'
+end
+
+group :development do
+  # Access an IRB console on exceptions page and /console in development
+  gem 'web-console', '~> 2.0.0.beta2'
+
+  # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
+  gem 'spring'
+
+  # Automated security checks
   gem 'brakeman'
-  gem 'railroady'
-  gem 'rspec-rails'
-  gem 'cucumber-rails', :require => false
+
+  # Time travel gem
+  gem 'timecop'
+
+  # Command line reference
+  gem 'cheat'
+
+  # Assorted generators
   gem 'nifty-generators'
+
+  # Class diagrams
+  gem 'rails-erd'
+  gem 'railroady'
+
+  # CoffeeScript source maps
+  gem 'coffee-rails-source-maps'
+end
+
+group :test do
+  # Use RSpec for tests
+  gem 'rspec-rails'
+
+  # Fixture replacement
   gem 'factory_girl_rails'
-  gem 'rack-test', require: 'rack/test'
+
+  # Lorem Ipsum
+  gem 'faker'
+
+  # Spec helpers
+  gem 'shoulda-matchers', require: false
+
+  # Code Climate integration
+  gem "codeclimate-test-reporter", require: false
+
+  # Coveralls integration
+  gem 'coveralls', require: false
 end
 
 group :production do
-  gem 'mysql2'
+  # Unicorn production server
   gem 'unicorn'
-  gem 'lograge', '~> 0.3.0'
+
+  # PostgreSQL production database
+  gem 'pg'
+
+  # Notify developers of Exceptions in production
+  gem 'exception_notification'
 end
