@@ -1,11 +1,10 @@
-# Copyright 2011-2013 Rice University. Licensed under the Affero General Public 
-# License version 3 or later.  See the COPYRIGHT file for details.
+# Copyright 2013-2014 Rice University. Licensed under the Affero General
+# Public License version 3 or later.  See the COPYRIGHT file for details.
 
 module ApplicationFooterHelper
 
   def copyright_text
-    year_range = "2013-#{Time.now.year}".sub(/\A(\d+)-\1\z/, '\1');
-    "Copyright &copy; #{year_range} #{COPYRIGHT_HOLDER}".html_safe
+    OpenStax::Utilities::Text.copyright(COPYRIGHT_START_YEAR, COPYRIGHT_HOLDER)
   end
 
 end
