@@ -1,13 +1,11 @@
 # Load the rails application
 require File.expand_path('../application', __FILE__)
 
-SITE_NAME = "OpenStax Exchange"
-COPYRIGHT_HOLDER = "Rice University"
+SITE_NAME = 'OpenStax Exchange'
+COPYRIGHT_HOLDER = 'Rice University'
 COPYRIGHT_START_YEAR = 2013
 
-require 'secret_settings'
 require 'controller_extensions'
-require 'utilities'
 require 'user'
 require 'app'
 require 'event'
