@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 user_event_controller_spec :input
 application_event_controller_spec :input, :create_multiple_choice
