@@ -1,10 +1,11 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe User do
 
   it 'adds user methods to relevant classes' do
     expect(ActiveRecord::Base).to respond_to(:acts_as_user)
-    expect(ActiveRecord::ConnectionAdapters::TableDefinition.new(:test)).to respond_to(:user)
+    expect(ActiveRecord::ConnectionAdapters::TableDefinition.new(
+      {}, :test, true, {})).to respond_to(:user)
     expect(ActiveRecord::Migration.new).to respond_to(:add_user_index)
     expect(ActionDispatch::Routing::Mapper.new(Exchange::Application.routes)).to(
       respond_to(:user_routes))
